VixAI-backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.0
requests==2.32.3
anyio==4.4.0
beautifulsoup4==4.12.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-ads==28.0.0
APScheduler==3.10.4
slowapi==0.1.9
